+++ hibernate/src/main/java/org/example/service/impl/TeacherServiceImpl.java
@@ -7,6 +7,8 @@ import org.example.service.TeacherService;
 import org.hibernate.Session;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 @AllArgsConstructor
 public class TeacherServiceImpl extends SessionService implements TeacherService {
@@ -15,7 +17,7 @@ public class TeacherServiceImpl extends SessionService implements TeacherService
     @Override
     public void save(TeacherEntity teacher) {
         Session session = getSession();
-        session.save(teacher);
+        session.persist(teacher);
         closeSession(session);
         session.close();
     }
@@ -23,18 +25,22 @@ public class TeacherServiceImpl extends SessionService implements TeacherService
     @Override
     public TeacherEntity get(Integer id) {
         Session session = getSession();
-        TeacherEntity teacher = session.get(TeacherEntity.class, id);
+        Optional<TeacherEntity> teacher = Optional.ofNullable( session.get(TeacherEntity.class, id));
         closeSession(session);
-        return teacher;
+        return teacher.orElse(new TeacherEntity());
     }
 
     @Override
-    public void delete(Integer id) {
+    public boolean delete(Integer id) {
         Session session = getSession();
-        TeacherEntity teacher = session.get(TeacherEntity.class,id);
-        session.delete(teacher);
+        boolean isDeleted = false;
+       Optional <TeacherEntity> teacherOptional = Optional.ofNullable(session.get(TeacherEntity.class,id));
+       if(teacherOptional.isPresent()){
+           TeacherEntity teacher = teacherOptional.get();
+           session.delete(teacher);
+           isDeleted=true;
+       }
         closeSession(session);
-
-
+       return isDeleted;
     }
 }
